Refactor trainer rank package API
